package/requirements.txt
boto3
pymysql
pandas
awswrangler
mysql-connector-python
